feat: improve loading state of hamster animation and use it in polkadot loading
