+++ src/components/auth/common/polkadot-connect/PolkadotConnectConfirmationContent.tsx
@@ -1,5 +1,6 @@
 import ProcessingHumster from '@/assets/graphics/processing-humster.png'
 import Button from '@/components/Button'
+import DynamicLoadedHamsterLoading from '@/components/DynamicLoadedHamsterLoading'
 import { AddProxyWrapper } from '@/services/subsocial/proxy/mutation'
 import { useSendEvent } from '@/stores/analytics'
 import { useMyAccount } from '@/stores/my-account'
@@ -27,12 +28,6 @@ export default function PolkadotConnectConfirmationContent({
   return (
     <div className='mt-2 flex flex-col gap-6'>
       <div className='flex w-full flex-col items-center gap-4'>
-        <Image
-          className='w-64 max-w-xs rounded-full'
-          priority
-          src={ProcessingHumster}
-          alt=''
-        />
         <AddProxyWrapper
           loadingUntilTxSuccess
           config={{
@@ -51,26 +46,41 @@ export default function PolkadotConnectConfirmationContent({
         >
           {({ isLoading, mutateAsync: addProxy }) => {
             return (
-              <Button
-                className='w-full'
-                size='lg'
-                onClick={async () => {
-                  const address = toSubsocialAddress(connectedWallet?.address)
-                  const signer = connectedWallet?.signer
-                  if (address && signer) {
-                    setIsProcessing(true)
-                    const shouldProceed = beforeAddProxy
-                      ? await beforeAddProxy?.()
-                      : true
-                    setIsProcessing(false)
-                    if (!shouldProceed) return
-                    addProxy(null)
-                  }
-                }}
-                isLoading={isLoading || isLoadingEnergy || isProcessing}
-              >
-                Confirm
-              </Button>
+              <>
+                <div className='mb-2'>
+                  {isLoading || isProcessing ? (
+                    <DynamicLoadedHamsterLoading />
+                  ) : (
+                    <Image
+                      className='w-64 max-w-xs rounded-full'
+                      priority
+                      src={ProcessingHumster}
+                      alt=''
+                    />
+                  )}
+                </div>
+
+                <Button
+                  className='w-full'
+                  size='lg'
+                  onClick={async () => {
+                    const address = toSubsocialAddress(connectedWallet?.address)
+                    const signer = connectedWallet?.signer
+                    if (address && signer) {
+                      setIsProcessing(true)
+                      const shouldProceed = beforeAddProxy
+                        ? await beforeAddProxy?.()
+                        : true
+                      setIsProcessing(false)
+                      if (!shouldProceed) return
+                      addProxy(null)
+                    }
+                  }}
+                  isLoading={isLoading || isLoadingEnergy || isProcessing}
+                >
+                  Confirm
+                </Button>
+              </>
             )
           }}
         </AddProxyWrapper>
